requirements.txt
playwright
psycopg2-binary
langchain
langchain-anthropic
langchain-community
langgraph
ddgs
opencc-python-reimplemented
openai
pgvector
